Restructure ipc files
