+++ tests/CMakeLists.txt
@@ -60,11 +60,11 @@ add_unit_test(utils/process)
 add_unit_test(components/command_line)
 add_unit_test(components/bar)
 add_unit_test(components/config_parser)
-add_unit_test(components/ipc_client)
 add_unit_test(drawtypes/label)
 add_unit_test(drawtypes/ramp)
 add_unit_test(drawtypes/iconset)
 add_unit_test(drawtypes/layouticonset)
+add_unit_test(ipc/client)
 add_unit_test(tags/parser)
 add_unit_test(tags/dispatch)
 add_unit_test(tags/action_context)
